Dyno v1.1 12PM error fixed. Auto set state on change. 1-fx sync button
Added function to auto-update timeslist when the section list is changed (added/removed), also changed conflict function to allow the end of one event to not collide with the start of another (before they couldn't be equal). Now ideally, I would want the site to auto-update the times list as well as run the permute function but for now sync does that. Sync used to have to be clickekd 2 since asynch wouldn't allow the updated permute to run after the set state in the same block.